+++ spec/rubocop/cop/rspec/capybara/current_path_expectation_spec.rb
@@ -26,4 +26,38 @@ RSpec.describe RuboCop::Cop::RSpec::Capybara::CurrentPathExpectation do
       current_path = WalkingRoute.last.path
     RUBY
   end
+
+  include_examples 'autocorrect',
+                   'expect(current_path).to eq expected_path',
+                   'expect(page).to have_current_path expected_path'
+
+  include_examples 'autocorrect',
+                   'expect(page.current_path).to eq(foo(bar).path)',
+                   'expect(page).to have_current_path(foo(bar).path)'
+
+  include_examples 'autocorrect',
+                   'expect(current_path).not_to eq expected_path',
+                   'expect(page).to have_no_current_path expected_path'
+
+  include_examples 'autocorrect',
+                   'expect(current_path).to_not eq expected_path',
+                   'expect(page).to have_no_current_path expected_path'
+
+  include_examples 'autocorrect',
+                   'expect(page.current_path).to match(/regexp/i)',
+                   'expect(page).to have_current_path(/regexp/i)'
+
+  include_examples 'autocorrect',
+                   'expect(page.current_path).to match("string/")',
+                   'expect(page).to have_current_path(/string\//)'
+
+  # Unsupported, no change.
+  include_examples 'autocorrect',
+                   'expect(page.current_path).to match(variable)',
+                   'expect(page.current_path).to match(variable)'
+
+  # Unsupported, no change.
+  include_examples 'autocorrect',
+                   'expect(page.current_path)',
+                   'expect(page.current_path)'
 end